--- js/GameState.js
@@ -1,17 +1,0 @@
-export const GameState = {
-  nickname: "",
-  selectedCharacter: null,
-  selectedStage: 1,
-  score: 0,
-  comboScore: 0,
-  comboCount: 0,
-  upgrades: [],
-  failedUpgrades: [],
-  reinforceChances: 3,
-  settings: {
-    bgm: true,
-    sfx: true,
-    cursor: true,
-  },
-  mapVisitedOnce: false
-};
