--- js/Settings.js
@@ -1,5 +1,0 @@
-import { GameState } from './GameState.js';
-
-export function toggleBGM() {
-  GameState.settings.bgm = !GameState.settings.bgm;
-}
