--- js/webTP.js
@@ -1,5 +1,0 @@
-import { showStartUI } from './GameStartUI.js';
-
-$(document).ready(() => {
-  showStartUI();
-});
