+++ webTP.html
@@ -8,6 +8,6 @@
 
 </head>
 <body>
-  <script type="module" src="./js/webTP.js"></script>
+  <script src="webTP.bundle.js"></script>
 </body>
 </html>
